# doc/openapi.yaml
openapi: "3.0.3"
info:
  title: "Inthub Backend API"
  version: "1.0.0"
  description: >
    この仕様書は、FastAPI によるバックエンド実装に対応する API を定義しています。
    各エンドポイントには、詳細な説明、想定されるエラー(400, 401, 404, 409, 500 など)も記載しています。
servers:
  - url: "https://inthub-fjfjehdsc4akamdg.japaneast-01.azurewebsites.net"
...
paths:
...
  /api/v1/rag/build-index:
    post:
      tags:
        - RAG
      summary: "RAGインデックス構築"
      description: >
        指定されたファイルIDに基づいてRAGインデックスを構築します。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildIndexRequest"
      responses:
        '200':
          description: "RAGインデックスの構築に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildIndexResponse"
        '400':
          description: "入力データが不正"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "認証失敗：無効または不足しているトークン"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/rag/delete-index:
    delete:
      tags:
        - RAG
      summary: "RAGインデックス削除"
      description: >
        現在のテナントに対するRAGインデックスを削除します。
      security:
        - BearerAuth: []
      responses:
        '200':
          description: "RAGインデックスの削除に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteIndexResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/rag/query:
    post:
      tags:
        - RAG
      summary: "RAGクエリ実行"
      description: >
        指定されたクエリに基づいてRAG検索を実行し、結果を返却します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagQueryRequest"
      responses:
        '200':
          description: "RAGクエリの実行に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagQueryResponse"
        '400':
          description: "入力データが不正"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BuildIndexRequest:
      type: object
      properties:
        file_id:
          type: string
      required:
        - file_id
    BuildIndexResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        indexSize:
          type: integer
      required:
        - status
        - message
        - indexSize
    DeleteIndexResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required:
        - status
        - message
    RagQueryRequest:
      type: object
      properties:
        query:
          type: string
        indexName:
          type: string
        topK:
          type: integer
          default: 5
      required:
        - query
        - indexName
    RagQueryResponse:
      type: object
      properties:
        status:
          type: string
        results:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - status
        - results
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        error_message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
      required:
        - error_code
        - error_message